--- Week1_Trapped_Ions/install_instruction.md
@@ -1,15 +1,0 @@
-# Installation Guide for Week1 Project
-
-## 0. Prerequisite
-All based on python >= 3.6.
-
-## 1. Install Julia
-Go to [julia download page](https://julialang.org/downloads/) and follow the step.
-Make sure your Julia binary directory to be added to `PATH`.
-
-## 2. Install Python Libraries & PyJulia
-Run the following on the terminal.
-```shell
-    $ pip install -r ./requirements.txt
-    $ python ./install_pyjulia.py
-```
