+++ Week1_Trapped_Ions/start_here.md
@@ -1,0 +1,43 @@
+# CDL Week 1 Cohort Project. Trapped Ions
+Author: QuNovaComputing, Inc.
+- - -
+
+## 1. Installation Guide
+
+### 1.0. Prerequisite
+All based on python >= 3.6.
+
+### 1.1. Install Julia
+Go to [julia download page](https://julialang.org/downloads/) and follow the step.
+Make sure your Julia binary directory to be added to `PATH`.
+
+### 1.2. Install Python Libraries & PyJulia
+Run the following on the terminal.
+```shell
+    $ pip install -r ./requirements.txt
+    $ python ./install_pyjulia.py
+```
+
+- - -
+
+## 2. Progress
+### 2.1. Technical Tasks
+- [x] Task 1 - Visualization of the distribution of measurement samples.
+- [x] Task 2 - Single bit-flip error
+- [x] Task 3 - Empirical distribution of $ p $
+- [x] Task 4 - Cross entropy benchmark fidelity for two-qubit gate error
+
+### 2.2. Buiseness Application
+- [x] Step 1
+- [x] Step 2
+- [x] Step 3
+- [x] Step 4
+
+### 2.3. Additional Challenges
+- [x] Additional Challenge 1 - Animation for Task 2.
+- [x] Additional Challenge 2 - Regeneration of fidelity from google paper.
+- [ ] Additional Challenge 3 - Using IonQ machine.
+
+- - -
+
+## 3.
